update socket operations to work with 64-bit

+++ inc/misc2.h
@@ -23,6 +23,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <winsock2.h>
 
 void wsabuf_get_buffer(uint32_t buffer_count, const WSABUF *buffers,
-    uint8_t **ptr, uint32_t *length);
+    uint8_t **ptr, uintptr_t *length);
 
 #endif
